pyproject.toml
[project]
name = "portfolio"
version = "0.0.1"
description = "Personal Website"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sphinx>=9.0.4",
    "sphinx-autobuild>=2025.8.25",
    "tibas-tt>=0.3.1",
]
